# api-docs/openapi.yaml
openapi: 3.1.0
info:
  title: Kayak Simulation Distributed Services API
  version: 1.0.0
  description: |
    API specification for the Kayak-style distributed system. The platform exposes REST
    endpoints for user management, listings search, bookings, billing, admin operations,
    analytics, and the FastAPI-based agentic recommendation service. All synchronous
    operations occur over HTTP while asynchronous updates are published on Kafka topics
    and relayed to clients via WebSockets.
  contact:
    name: Kayak Simulation Platform Team
    email: eng@kayak-sim.local
servers:
  - url: https://api.kayak-sim.prod
    description: Production API Gateway
  - url: https://api.staging.kayak-sim.local
    description: Staging API Gateway
  - url: http://localhost:8080
    description: Local development gateway
...
paths:
...
  /api/v1/payments:
    get:
      tags: [Billing]
      summary: List payments
      operationId: listPayments
      parameters:
        - name: userId
          in: query
          schema:
            type: string
            format: uuid
        - name: bookingId
          in: query
          schema:
            type: string
            format: uuid
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/PaymentStatus'
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PageSizeParam'
      responses:
        '200':
          description: Payment records.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentListResponse'
    post:
      tags: [Billing]
      summary: Create payment
      description: Initiate a payment for a booking. Performs idempotency on client token.
      operationId: createPayment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PaymentCreateRequest'
      responses:
        '201':
          description: Payment initiated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Payment'
        '400':
          $ref: '#/components/responses/ValidationError'
        '402':
          description: Payment required / card declined.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Duplicate payment attempt detected for the booking.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/payments/{paymentId}:
    parameters:
      - name: paymentId
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      tags: [Billing]
      summary: Get payment
      operationId: getPayment
      responses:
        '200':
          description: Payment details.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Payment'
        '404':
          $ref: '#/components/responses/NotFoundError'
  /api/v1/payments/{paymentId}/refunds:
    post:
      tags: [Billing]
      summary: Issue refund
      operationId: refundPayment
      parameters:
        - name: paymentId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefundRequest'
      responses:
        '202':
          description: Refund accepted for processing.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Payment'
        '404':
          $ref: '#/components/responses/NotFoundError'
...
  /api/v1/admin/reports/revenue:
    get:
      tags: [Admin, Analytics]
      summary: Revenue report
      operationId: getRevenueReport
      security:
        - bearerAuth: []
      parameters:
        - name: year
          in: query
          required: true
          schema:
            type: integer
            minimum: 2000
        - name: groupBy
          in: query
          schema:
            type: string
            enum: [city, provider, property, month]
            default: city
        - name: bookingType
          in: query
          schema:
            $ref: '#/components/schemas/BookingType'
      responses:
        '200':
          description: Aggregated revenue metrics.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RevenueReportResponse'
        '404':
          $ref: '#/components/responses/NotFoundError'
  /api/v1/admin/reports/providers:
    get:
      tags: [Admin, Analytics]
      summary: Top providers report
      operationId: getTopProviders
      security:
        - bearerAuth: []
      parameters:
        - name: month
          in: query
          required: true
          schema:
            type: string
            pattern: '^(19|20)\\d{2}-(0[1-9]|1[0-2])$'
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 10
      responses:
        '200':
          description: Provider performance snapshot.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProviderPerformanceResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  parameters:
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
      description: Page number for pagination (1-indexed).
    PageSizeParam:
      name: pageSize
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 25
      description: Number of items to return per page.
...
  responses:
    ValidationError:
      description: Input validation failed; see details array.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationErrorResponse'
    NotFoundError:
      description: Resource not found.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
    PaginationMetadata:
      type: object
      required: [page, pageSize, totalItems, totalPages]
      properties:
        page:
          type: integer
          minimum: 1
        pageSize:
          type: integer
          minimum: 1
          maximum: 200
        totalItems:
          type: integer
          minimum: 0
        totalPages:
          type: integer
          minimum: 0
    ErrorResponse:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          description: Machine-readable error code.
        message:
          type: string
          description: Human-readable error description.
        correlationId:
          type: string
          format: uuid
          description: Trace ID for support teams.
        details:
          type: array
          items:
            type: string
          description: Additional error context.
    ValidationErrorResponse:
      allOf:
        - $ref: '#/components/schemas/ErrorResponse'
        - type: object
          properties:
            errors:
              type: array
              description: Field level validation failures.
              items:
                type: object
                required: [field, message]
                properties:
                  field:
                    type: string
                  message:
                    type: string
                  rejectedValue:
                    description: The value that failed validation.
    Address:
      type: object
      required: [line1, city, state, zipCode]
      properties:
        line1:
          type: string
          maxLength: 255
        line2:
          type: string
          maxLength: 255
        city:
          type: string
          maxLength: 120
        state:
          $ref: '#/components/schemas/USStateCode'
        zipCode:
          type: string
          pattern: '^\\d{5}(-\\d{4})?$'
          description: Supports '#####' or '#####-####'.
    PaymentInstrument:
      type: object
      required: [paymentType, cardLast4]
      properties:
        paymentType:
          type: string
          enum: [credit_card, debit_card, paypal, apple_pay, google_pay]
        cardBrand:
          type: string
          example: VISA
        cardLast4:
          type: string
          pattern: '^\\d{4}$'
        expirationMonth:
          type: integer
          minimum: 1
          maximum: 12
        expirationYear:
          type: integer
          minimum: 2024
        billingAddress:
          $ref: '#/components/schemas/Address'
...
    Money:
      type: object
      required: [amount, currency]
      properties:
        amount:
          type: number
          format: float
          minimum: 0
        currency:
          type: string
          pattern: '^[A-Z]{3}$'
          example: USD
    BookingType:
      type: string
      enum: [flight, hotel, car]
...
    PaymentStatus:
      type: string
      enum: [PENDING, AUTHORIZED, SUCCEEDED, FAILED, REFUNDED]
    Payment:
      type: object
      required:
        - id
        - bookingId
        - userId
        - status
        - amount
        - createdAt
      properties:
        id:
          type: string
          format: uuid
        bookingId:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        status:
          $ref: '#/components/schemas/PaymentStatus'
        amount:
          $ref: '#/components/schemas/Money'
        paymentMethod:
          $ref: '#/components/schemas/PaymentInstrument'
        transactionReference:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        invoiceUrl:
          type: string
          format: uri
    PaymentCreateRequest:
      type: object
      required:
        - bookingId
        - userId
        - amount
        - paymentMethodId
      properties:
        bookingId:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        amount:
          $ref: '#/components/schemas/Money'
        paymentMethodId:
          type: string
          format: uuid
        idempotencyKey:
          type: string
          maxLength: 50
        metadata:
          type: object
          additionalProperties: true
    PaymentListResponse:
      type: object
      required: [items, pagination]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Payment'
        pagination:
          $ref: '#/components/schemas/PaginationMetadata'
    RefundRequest:
      type: object
      properties:
        amount:
          $ref: '#/components/schemas/Money'
        reason:
          type: string
...
    RevenueReportResponse:
      type: object
      required: [generatedAt, granularity, items]
      properties:
        generatedAt:
          type: string
          format: date-time
        granularity:
          type: string
          enum: [city, provider, property, month]
        currency:
          type: string
          pattern: '^[A-Z]{3}$'
          default: USD
        items:
          type: array
          items:
            type: object
            required: [key, revenue, bookings]
            properties:
              key:
                type: string
              revenue:
                $ref: '#/components/schemas/Money'
              bookings:
                type: integer
                minimum: 0
              averageOrderValue:
                $ref: '#/components/schemas/Money'
    ProviderPerformanceResponse:
      type: object
      required: [generatedAt, providers]
      properties:
        generatedAt:
          type: string
          format: date-time
        providers:
          type: array
          items:
            type: object
            required: [providerId, providerName, unitsSold, revenue]
            properties:
              providerId:
                type: string
              providerName:
                type: string
              unitsSold:
                type: integer
                minimum: 0
              revenue:
                $ref: '#/components/schemas/Money'
              month:
                type: string
                pattern: '^(19|20)\\d{2}-(0[1-9]|1[0-2])$'
...
    USStateCode:
      type: string
      enum:
        - AL
        - AK
        - AZ
        - AR
        - CA
        - CO
        - CT
        - DE
        - FL
        - GA
        - HI
        - ID
        - IL
        - IN
        - IA
        - KS
        - KY
        - LA
        - ME
        - MD
        - MA
        - MI
        - MN
        - MS
        - MO
        - MT
        - NE
        - NV
        - NH
        - NJ
        - NM
        - NY
        - NC
        - ND
        - OH
        - OK
        - OR
        - PA
        - RI
        - SC
        - SD
        - TN
        - TX
        - UT
        - VT
        - VA
        - WA
        - WV
        - WI
        - WY
      description: Two-letter postal abbreviation for US states.
...